Update Embedding dimensions and package references
Modified `Dimensions` property in `BigANNBinaryEmbeddingOnlyScearioBase.cs` to cast `EmbeddingDimensions` to an integer. Updated `Microsoft.Azure.Cosmos` package to `3.46.1` and added `Newtonsoft.Json` package version `13.0.3` in `VectorIndexScenarioSuite.csproj`.

+++ BigANNBinaryEmbeddingOnlyScearioBase.cs
@@ -86,7 +86,7 @@ namespace VectorIndexScenarioSuite
                         Path = this.EmbeddingPath,
                         DataType = this.EmbeddingDataType,
                         DistanceFunction = this.EmbeddingDistanceFunction,
-                        Dimensions = this.EmbeddingDimensions,
+                        Dimensions = (int) this.EmbeddingDimensions,
                     }
                 })),
                 IndexingPolicy = new IndexingPolicy()
